requirements.txt
qrcode
Pillow
PyYAML
